Created QZXingEncoderConfig struct to hold the configuration of the encode operation. It has been extracted to a struct since the parameters are increasing.
Created overloaded function for encodeData to use the QZXingEncoderConfig. The old implementation of encodeData now translates its arguments into a QZXingEncoderConfig and calls the overloaded function.

Updated regex in QZXingImageProvider to be aware of the new configuration of border.

Updated BarcodeEncoder example UI to allow the control of the border option.